smart passenger update
