+++ multimodal-ui/src/app/interfaces/tags.model.ts
@@ -1,0 +1,13 @@
+export interface Tagged {
+  tags: string[];
+}
+
+export function isTagged(value: unknown): value is Tagged {
+  return (
+    typeof value === 'object' &&
+    value !== null &&
+    'tags' in value &&
+    Array.isArray((value as Tagged).tags) &&
+    (value as Tagged).tags.every((tag) => typeof tag === 'string')
+  );
+}
