+++ python/multimodalsim_viewer/server/http_routes.py
@@ -7,9 +7,7 @@ import zipfile
 from flask import Blueprint, jsonify, request, send_file
 
 from multimodalsim_viewer.common.utils import get_input_data_directory_path
-from multimodalsim_viewer.server.simulation_visualization_data_model import (
-    SimulationVisualizationDataManager,
-)
+from multimodalsim_viewer.server.data_manager import SimulationVisualizationDataManager
 
 http_routes = Blueprint("http_routes", __name__)
